requirements.txt
openai==1.12.0
neo4j==5.15.0
beautifulsoup4==4.12.2
requests==2.31.0
tiktoken==0.5.2
flask==3.0.0
python-dotenv==1.0.0